fit binned light curves for t0, rprs, and systematics
heavily modified an exoplanet transit fitting script by Dr. Munazza Alam to loop through an xarray of spectroscopic light curves and fit each of those for mid transit time, planet radius in stellar radii, and systematics (such as slope). nrs2 fails in some cases for a reason i was unable to discern, but minimally enough that i could move on with the project as it is